JDP230701-22 - FIX + moving Exceptions
