+++ src/main/java/com/kodilla/ecommercee/controller/CartController.java
@@ -3,11 +3,7 @@ package com.kodilla.ecommercee.controller;
 import com.kodilla.ecommercee.domain.Cart;
 import com.kodilla.ecommercee.domain.CartDto;
 import com.kodilla.ecommercee.domain.ProductDto;
-import com.kodilla.ecommercee.domain.ProductNotFoundException;
-import com.kodilla.ecommercee.exception.AllreadyActiveCartException;
-import com.kodilla.ecommercee.exception.CartNotFoundException;
-import com.kodilla.ecommercee.exception.NotActiveCartException;
-import com.kodilla.ecommercee.exception.NotAutorizedUserException;
+import com.kodilla.ecommercee.exception.*;
 import com.kodilla.ecommercee.mapper.CartMapper;
 import com.kodilla.ecommercee.mapper.ProductMapper;
 import com.kodilla.ecommercee.service.CartDbService;
