+++ src/main/java/com/kodilla/ecommercee/mapper/GroupProductMapper.java
@@ -2,17 +2,24 @@ package com.kodilla.ecommercee.mapper;
 
 import com.kodilla.ecommercee.domain.GroupProduct;
 import com.kodilla.ecommercee.domain.GroupProductDto;
+import com.kodilla.ecommercee.domain.Product;
+import com.kodilla.ecommercee.domain.ProductDto;
+import com.kodilla.ecommercee.exception.GroupProductNotFoundException;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
+import java.util.List;
 
 @Service
 public class GroupProductMapper {
+    @Autowired
+    private ProductMapper productMapper;
 
     public GroupProduct mapToGroupProduct(final GroupProductDto groupProductDto) {
         return new GroupProduct(groupProductDto.getId(),
                 groupProductDto.getName(),
-                groupProductDto.getProducts() != null ? groupProductDto.getProducts() : new ArrayList<>()
+                getProductList(groupProductDto.getProducts())
         );
     }
 
@@ -20,7 +27,21 @@ public class GroupProductMapper {
         return new GroupProductDto(
                 groupProduct.getId(),
                 groupProduct.getName(),
-                groupProduct.getProducts()
+                productMapper.mapToProductDtoList(groupProduct.getProducts())
         );
+    }
+    private List<Product> getProductList(final List<ProductDto> productsDto) {
+        if (productsDto == null) {
+            return new ArrayList<>();
+        }
+
+        return productsDto.stream()
+                .map(productDto -> {
+                    try {
+                        return productMapper.mapToProduct(productDto);
+                    } catch (GroupProductNotFoundException e) {
+                        throw new RuntimeException(e);
+                    }
+                }).toList();
     }
 }
